Ginkgo Gauss Seidel halo kernels and projecting in single precision (#12)
* Fix straggling uses of mat_scalar_type.

* Ginkgo smoother with halo kernel.

* Matrix sendBuffer needs to be of type local_scalar_t, as it is used by the vectors.

* Add profiling ranges for ginkgo_smoother_, etc.

* Cleanup estimate_run_time definition.

* GinkgoSmoother fixes.

* Try using matrix local number of rows in Ginkgo sizes.

* ell_multicolor_gs --> multicolor_gs in profiling ranges.

* Perform all FGS_HALO colors in one kernel launch in GinkgoSmoother.

* Workaround to perform projection in single-precision when instantiating AMP<double>.

* A few more profiling ranges in the Ginkgo interface.

* Keep MGData in halo_scalar_t.

* Use default (larger) performance counters when using GinkgoSmoother for the first pre-smoother application.

* Use initial_guess_mode::provided (instead of zero default). Changing from 2 FGS iter to 1 in ginkgo/core increased the number of GMRES iter here, but the different initial guess seems to help.

* Try getting AMP tolerance from file; list bin statistics.

* Instantiate Ginkgo executor with compute stream.

* Update src/GenerateNonsymCoarseProblem.cpp


* Pass HPGMP_gen_opt structure down the stack to enable getting amp_tol in the GinkgoMatrix constructor.

* HPGMP_gen_opt fixes for other builds (including reference).

* [skip ci] pre-commmit --> pre-commit.

---------

+++ src/GMRESData.cpp
@@ -3,4 +3,4 @@
 // Available instantiations
 template class GMRESData<double, double, double>;
 template class GMRESData<float, float, float>;
-template class GMRESData<double, float, double>;
+template class GMRESData<double, float, float>;
